not all search is uct
