--- uct/__init__.py
@@ -1,3 +1,0 @@
-from uct.uct import NeuralNet
-from uct.uct import UCT_search
-# from uct.util import softmax, temp_softmax
